+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # CHANGELOG
+### Version 2.1.1
+
+
 ### Version 2.0.9
 - ADD: Purity percentage in cart
 - FIX: Show an error in impact and pyramid view if no ingredients